[#1087] Made a temporary split of logic of Details into separate component.
