+++ frontend/src/CommitPanel/details/Diff.tsx
@@ -1,12 +1,12 @@
 import * as React from 'react';
 
-import DiffPanel from '../Commits/DiffPanel.react';
+import DiffPanel from '../../Commits/DiffPanel.react';
 
 interface CommitPanelDetailsProps extends React.Props<JSX.Element> {
   diff: string;
 }
 
-export default class CommitPanelDetails extends React.Component<CommitPanelDetailsProps, {}> {
+export default class Diff extends React.Component<CommitPanelDetailsProps, {}> {
 
   render() {
     return <DiffPanel diff={this.props.diff} />;
